specific functions

+++ MeetingsWebAPI/Models/Location.cs
@@ -8,8 +8,8 @@ namespace MeetingsWebAPI.Models
     public class Location
     {
         public int LocationId { get; set; }
-        public double Lattitude { get; set; }
-        public double Longitude { get; set; }
+        public int Lattitude { get; set; }
+        public int Longitude { get; set; }
 
         public ICollection<Person> Persons { get; set; }
     }
